refactor(asset): moving meta implementation to schema
