+++ modules/metas/internal/simulator/genesis.go
@@ -16,9 +16,10 @@ import (
 	"github.com/AssetMantle/modules/modules/metas/internal/parameters"
 	"github.com/AssetMantle/modules/modules/metas/internal/parameters/dummy"
 	"github.com/AssetMantle/modules/schema/data"
-	"github.com/AssetMantle/modules/schema/data/base"
+	baseData "github.com/AssetMantle/modules/schema/data/base"
 	"github.com/AssetMantle/modules/schema/helpers"
 	baseHelpers "github.com/AssetMantle/modules/schema/helpers/base"
+	"github.com/AssetMantle/modules/schema/mappables/base"
 	parameters2 "github.com/AssetMantle/modules/schema/parameters"
 	baseSimulation "github.com/AssetMantle/modules/simulation/schema/types/base"
 )
@@ -31,13 +32,13 @@ func (simulator) RandomizedGenesisState(simulationState *module.SimulationState)
 		dummy.ID.String(),
 		&Data,
 		simulationState.Rand,
-		func(rand *rand.Rand) { Data = base.NewDecData(sdkTypes.NewDecWithPrec(int64(rand.Intn(99)), 2)) },
+		func(rand *rand.Rand) { Data = baseData.NewDecData(sdkTypes.NewDecWithPrec(int64(rand.Intn(99)), 2)) },
 	)
 
 	mappableList := make([]helpers.Mappable, simulationState.Rand.Intn(99))
 
 	for i := range mappableList {
-		mappableList[i] = mappable.NewMeta(baseSimulation.GenerateRandomData(simulationState.Rand))
+		mappableList[i] = mappable.NewMappable(base.NewMeta(baseSimulation.GenerateRandomData(simulationState.Rand)))
 	}
 
 	genesisState := baseHelpers.NewGenesis(key.Prototype, mappable.Prototype, nil, parameters.Prototype().GetList()).Initialize(mappableList, []parameters2.Parameter{dummy.Parameter.Mutate(Data)})
